Add incomplete_package_import option to speed reposync up
Considerable time is taken by spacewalk-repo-sync and inter-server sync
due to database insertions in tables with a high number of rows.

Three of these (rhnPackageFile, rhnPackageChangeLogRec and
rhnPackageChangeLogData) only contain informational data that can be skipped
in order to get faster repo-sync times, at the expense of not finding this
data in the Web UI.

This patch adds an rhn.conf flag (incomplete_package_import = 1) that will result
in spacewalk-repo-sync and inter-server sync skipping those tables with
noticeable performance gains.

This is particularly useful in tests and demos, and is not recommended in
production.